Adding a decorator for attaching a prefix to the log context
This will make any updates within a method call have a prefix +
definable separator. Note that it will only raise errors at runtime.
Also, I didn't bother with adding the ability to clear log context in
the case of an exception if a user wants but this would be relatively
trivial to do.